requirements.txt
pip 
numpy 
pillow 
torch 
torchvision 
torchaudio